main: improve api socket bind
If the grout process is launched a second time, then socket api bind() failed
and the process unlink socket file before exiting, making the first running
process unreachable.

If the grout process crash, it won't remove api socket file, and the next
invocation will fail at socket api bind().

Change api_socket_start() to check, if bind() fails, whether there's a
listening process behind the scene:

* If there is, exit without removing socket api file
* If there is not, then remove file and resume bind.

+++ main/main.c
@@ -147,6 +147,8 @@ end:
 	return 0;
 }
 
+#define EXIT_ALREADY_RUNNING 2
+
 int main(int argc, char **argv) {
 	struct event_base *ev_base = NULL;
 	int ret = EXIT_FAILURE;
@@ -183,6 +185,8 @@ int main(int argc, char **argv) {
 	modules_init(ev_base);
 
 	if (api_socket_start(ev_base) < 0) {
+		if (errno == EADDRINUSE)
+			ret = EXIT_ALREADY_RUNNING;
 		err = errno;
 		goto shutdown;
 	}
@@ -211,7 +215,8 @@ shutdown:
 		modules_fini(ev_base);
 		event_base_free(ev_base);
 	}
-	unlink(args.api_sock_path);
+	if (ret != EXIT_ALREADY_RUNNING)
+		unlink(args.api_sock_path);
 	libevent_global_shutdown();
 dpdk_stop:
 	dpdk_fini();
